Add `shouldFocusInput` option [Fix #1541]
This adds a `shouldFocusInput` option that determines whether or
not the focusser should be automatically focused.  By default, the
focusser will always be focused, unless `minimumResultsForSearch`
is less than zero.

`close` on single selects took an option `params` argument which
at one point was used to tell the focusser to not automatically
focus.  This was added at one point when tweaking the mask [1]
and later reverted to fix a bug that came as a result [2].  The
leftovers of this code has been removed in this commit.

[1]: https://github.com/ivaynberg/select2/commit/e162a4802cddfb95fdef8793fc83fc07897650ec
[2]: https://github.com/ivaynberg/select2/commit/9bc68f089e4d81e0dd0e8de943071ae36a3394cd